refactor : submit section 추상화 제거
