+++ app/join/funnels/step5/components/InputForm.client.tsx
@@ -5,10 +5,26 @@ import SubmitSection from './SubmitSection.client';
 import { useSignUpMutation } from '@/apis/auth';
 import { useJoinContext } from '@/app/join/components/JoinContext';
 import { SignUpState } from '@/app/join/type';
+import { BottomFixedButton } from '@/components/common/Button';
 import { personalityList } from '@/constants/personalityList';
 
+const DUMMY_SIGN_UP_DATA = {
+  phoneNumber: '010-5728-9353',
+  imageUrl:
+    'https://gloddy.s3.ap-northeast-2.amazonaws.com/file/87d8b6c4-fcda-4588-8334-b3ca96e635a0.png',
+  schoolInfo: {
+    school: '가천대학교',
+    email: 'gueit214@gachon.ac.kr',
+    certifiedStudent: true,
+  },
+  nickname: 'string',
+  birth: '2023-07-22',
+  gender: 'MAIL',
+  personalities: ['OUTGOING'],
+};
+
 export default function InputForm() {
-  const { handleSubmit } = useJoinContext();
+  const { handleSubmit, getValues } = useJoinContext();
   const { mutate: mutateSignUp } = useSignUpMutation();
   const onSubmit = async (data: SignUpState) => {
     const { certificateEmailNumber, certificateNumber, birth, gender, personalityIdList, ...rest } =
@@ -26,7 +42,11 @@ export default function InputForm() {
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
       <PersonalitySection />
-      <SubmitSection />
+      <BottomFixedButton
+        text="완료"
+        disabled={getValues().personalityIdList.length === 0}
+        type="submit"
+      />
     </form>
   );
 }
